Support per-sensor/output/service logging
This allows for granular customisation e.g.

- To get DEBUG logs, but only for one sensor of concern.
- To reduce the log volume if a sensor polls frequently.

The new approach also has some general advantages:

- The log structure for sensors is the same as outputs.

- An arbitrary __str__ overload is no longer required.

- The log structure is more compatible with logs from
other packages and future non-thread-based sources  by
not depending on the thread names for grouping.

The new approach has a disadvantage in that a developer
now needs to decide how to access the logger:

- "self.logger" should be used wherever possible.

- "get_logger()" tries to find an appropriate logger for
the current thread - if "self.logger" not available.