Generalise stream filters into arbitrary functions
Filters are just one kind of mutation on a Stream:

- Averaging readings over a window (fixed or moving).
- Changing the attribute of each reading (no use case).
- Grouping readings together (also no use case yet).

Averaging will be useful in upcoming work:

- To limit the amount of output data when a sensor
has to be polled very frequently (but not every sample
is wanted for storage purposes).

- To downsample data for longterm storage.

Filters and Windows are intentionally incomplete and
ready for subclasses to extend their functionality,
which can then be added to Stream for chaining e.g.

    Stream.special_avg = lambda size: WindowSpecialAverage(size)
    ...
    MockSensor().stream.special_avg(size=3)...

A downside of the name "function" is it's also a term
in Python. I couldn't think of a clearer name, though:

- "map" doesn't cover filters
- "transform" doesn't cover filters (really)
- "mutation" implies non-functional behaviour

A function in "Snsary" really is a function - just a
specific kind - so I don't think it will be a problem
in practice. Could always deprecate if it is.