Add batch GraphQL Output for DataStax Astra DB
This is intended to be used for longterm storage. A rename feature
allows for storing aggregations in the same table, noting that the
DB and the GraphQL query API can't aggregate automatically.

The datasource plugin for Cassandra doesn't currently work with V8
Grafana, but the GraphQL plugin does. By the time the data has built
up the situation will hopefully have improved.

Initial testing with the GraphQL plugin didn't raise any concerns with
storing aggregations together in the same table, which is certainly
easier than supporting multiple tables for the GraphQL output.

Given the complexity of the schema request, it was tempting to use the
Betamax library as a VCR alternative to manually storing the response
data. I decided it wasn't worth it for a single request.