Rewrite windows to support multiple output stats
This supports more advanced aggregations with multiple values, such
as the new WindowSummary function. Although this could be achieved
with individual functions and multiple chains in the calling code,
it's simpler to compute them all in one place.

Given that windows may be applied over long periods, it makes sense
to allow multiple units for the period argument.

The richer behaviour of Windows isn't ideal for Filter, which could
potentially be simplified in future with a suitable parent class to
abstract the array output. Given the additional complexity, this was
a good opportunity to remove the unused Stream.filter.