Simplify filtering and reduce aggregation load
It's no longer important to try and limit the amoumt of data going
to InfluxDB in order to cap longterm usage - longterm data is now
going to DataStax. A compromise of removing the filters is there's
now too much data going to DataStax, so this removes the sub-minute
portion of the data - per-minute should suffice longterm.

The simplified system should now be easier to extend with e.g. a
feedback loop for the new SGP30 sensor.