Prepare Reading model to support persistence
This model can't link to complex objects like the source / sensor,
as these can't be restored / persisted.

This is also a good opportunity to shorten the verbose "timestamp"
argument to match the naming used elsewhere - "timestamp" is used
by the datetime module to mean the number of seconds since epoch,
so the new name shouldn't be too ambiguous.