Allow logging from an anonymous source
This supports logging where the source thread has not yet or will
never identify itself by calling get_logger with a name argument.

While we could potentially fix some issues with preemptive calls
to self.logger - e.g. when a service starts (and thus has a thread
ID) - such techniques are not clearly comprehensive.

Storage will be initialised on first use, which means there's no
single logger that's appropriate - a good case for allowing anon
loggers to be created when a better name is not known.