Add framework for optional persistent storage
This was originally implemented using the shelve module but ended
up lower level - using pickle directly - in order to provide a TTL
feature via TTLCache. cachetools also provides a memoising @cached
decorator, which makes sense to use instead of functools.cache.

Note: expiry is not tested as it's done implicitly when a value is
next set on a TTLCache. It shouldn't be an issue if expired items
continue to exist a little longer than the TTL, provided they get
cleared eventually - the point is to prevent long term leakage.