Fix double UTC conversion for GraphQL output
This is the only output using the reading.datetime function, which
takes a timestamp and attempts to reconstitute it as a datetime.

Unfortunately I misunderstood utcfromtimestamp - I thought the UTC
naming referred to the timezone of the timestamp, but this doesn't
make sense as timestamps do not have a timezone - they are just the
number of seconds since UNIX epoch, without any location info.

What the function actually does is load the timestamp as a datetime
in the local timezone - this is correct - and then convert it to a
naive datetime in UTC. Without careful handling to reflect it's in
UTC, this effectively subtracts an hour from the original time e.g.

    datetime.utcfromtimestamp(1650785427).timestamp()
    1650781827.0

fromtimestamp still naively conveys the loaded timestamp, but in the
local timezone like all the others.