Switch all times to be in the local zone
"utcnow" returns a naive timestamp, which effectively subtracts an
hour from the local time, which is what we're trying to represent.

    datetime.fromtimestamp(datetime.utcnow().timestamp())
    datetime.datetime(2022, 4, 24, 8, 8, 29, 576299)

    datetime.fromtimestamp(datetime.now().timestamp())
    datetime.datetime(2022, 4, 24, 9, 9, 5, 365402)

Using "now" is correct. Because reasoning about naive times can be
so confusing, this also makes all of them timezone-aware e.g.

    datetime.utcnow().isoformat()
    '2022-04-26T07:13:24.259928'

    datetime.now().astimezone().isoformat()
    '2022-04-26T08:13:58.523714+01:00'

Using "astimezone" converts the datetime object to the specified
timezone by interpreting it in the implicit local timezone.

Tests now set explicit timestamps or zone-aware datetimes to make it
clear what actual time is being tested. Unfortunately freeze_gun is
not a reliable tool for this [^1][^2] so this commit replaces it.

In the case of the Octopus sensor, we are effectively converting a
date to a time so we also need to lock the system timezone to ensure
test stability. time-machine needs Python 3.9+ to do this [^3].

[^1]: https://github.com/spulec/freezegun/issues/348
[^2]: https://github.com/spulec/freezegun/issues/299
[^3]: https://github.com/adamchainz/time-machine#timezone-mocking