Switch to BigQuery Storage API (free tier)
While the original package was simpler (albeit version-limited), it
relies on a legacy API and isn't available in the Google Cloud free
tier [^1]. The BigQuery Storage library is relatively low level.

Using the default stream to send data avoids additional setup and is
pre-optimised for continuous sending of data [^2].

The BigQuery Storage package depends on GRPCIO, which seems to have
issues installing on Raspberry PI in the latest versions [^3]. It's
unclear if different versions are only meant to work with certain OS
ranges, or if there's a bug in the newer versions.

Although mocking the GRPCIO requests may be impractical, it does seem
to be possible [^4]. I don't think the value of functional testing is
sufficient to merit further investigation though.

[^1]: https://cloud.google.com/bigquery/pricing#data_ingestion_pricing
[^2]: https://cloud.google.com/bigquery/docs/write-api
[^3]: https://github.com/grpc/grpc/issues/30039#issuecomment-1186173717
[^4]: https://github.com/tokopedia/gripmock