Add black linter to standardise formatting
This removes unnecessary formatting decisions:

- Quotes are all double quotes
- Newlines are used deterministicaly

In the latter case, the removal of newlines has exposed a few areas
of code that relied on formatting to manage complexity, the common
example being nested comprehensions; the next commits will look at
rewriting the affected areas to be less dense.

Since black formats automatically, it seems reasonable to let it run
on all files, even though some are autogenerated.