Fix isort misreading "snsary" as third party
This was causing a problem when running isort from another virtual
environment - e.g. within Vim - as it could only see "snsary" as an
installed package, even though it was installed from local source.

Telling isort that "snsary" should always be a first-party package
stops it grouping the imports along with third-party ones.