Move poller exception handling up to poller
This is more intuitive as the need to handle exceptions is in order
to keep the poller process running until it's stopped.

This change also makes it possible to decorate the "tick" method with
new tracing to (optionally) monitor exceptions from it.

Note that the exception handling at either level is covered as part of
the main integration tests.