Use PyPMS reader instead of low-level serial cmds
This is viable since [^1].

There are some minor differences to note:

- SensorReader._cmd uses Serial.flush(), but this shouldn't be an
issue in newer versions of Python [^2].

- SensorReader._cmd reads all available input into the buffer and
only calls Serial.reset after a "_read", or if the buffer turns out
to be invalid. This is less robust than just calling Serial.reset
before each command, but it seems to work.

- SensorReader.open does an additional check to verify the sensor.

[^1]: https://github.com/avaldebe/PyPMS/pull/33
[^2]: https://github.com/python/cpython/issues/97001